+++ react-app/src/components/Comment/CommentContext.tsx
@@ -1,21 +1,14 @@
 import WinnerCard from '@/components/WinnerCard/WinnerCard';
-import { useEffect, useState } from 'react';
-import { getLatestPollIds } from '@/apis/poll/getPollLatest';
-import { CommentResponse } from '@/apis/comment/getComment';
-import { getComments } from '@/apis/comment/getAllComments';
 import CommentCard from '@/components/Comment/Comment';
 import CharacterCard from '@/components/Character/Character';
 import useIsMobile from '@/hook/useIsMobile';
 import CommentInputField from '@/components/Comment/CommentInputField';
-import { postComment } from '@/apis/comment/postComment';
-import { useRef } from 'react';
 import { useSelector } from 'react-redux';
 import { RootState } from '@/store';
 import { IMAGES } from '@/constants/imagePath';
-import {
-  getVoteResultByPollId,
-  getVoteResultByPollIdResponse,
-} from '@/apis/vote/getVoteResultByPollId';
+import { useComment } from '@/hook/useComment';
+import { useEffect } from 'react';
+import Loading from '../Loading/Loading';
 
 interface CommentProps {
   setIsLoading: (loading: boolean) => void;
@@ -23,153 +16,54 @@ interface CommentProps {
 }
 
 const CommentContext = ({ setIsLoading, onError }: CommentProps) => {
-  // 사용자 정보
   const user = useSelector((state: RootState) => state.auth.user)!;
   const isMobile = useIsMobile();
 
-  // 결과 조회
-  const [results, setResults] = useState<getVoteResultByPollIdResponse[]>([]);
-  const [pollIds, setPollIds] = useState<number[]>([]);
-
-  // 댓글 조회
-  const [comments, setComments] = useState<CommentResponse[]>([]);
-  // 댓글 입력창 상태 추가
-  const [commentText, setCommentText] = useState('');
-
-  // 댓글과 결과 모두 로딩 완료 시 false
-  const [resultsLoading, setResultsLoading] = useState(true);
-  const [commentsLoading, setCommentsLoading] = useState(true);
-
-  // 댓글 맨 아래로 내리기
-  const commentsEndRef = useRef<HTMLDivElement>(null);
-
-  // 최신 pollIds 불러오기
-  useEffect(() => {
-    const fetchAllResults = async () => {
-      try {
-        const res = await getLatestPollIds();
-        setPollIds(res ?? []);
-      } catch (error) {
-        console.error('투표 결과 불러오기 실패 : ', error);
-        setPollIds([]);
-        onError?.(error);
-      }
-    };
-    fetchAllResults();
-  }, [onError]);
-
-  // pollIds가 갱신되면 결과 가져오기
-  useEffect(() => {
-    const fetchResultsByPollIds = async () => {
-      try {
-        if (pollIds.length === 0) {
-          setResults([]);
-          setResultsLoading(false);
-          return;
-        }
-
-        setResultsLoading(true);
-        const res = await Promise.all(pollIds.map((pollId) => getVoteResultByPollId({ pollId })));
-        setResults(res);
-      } catch (error) {
-        console.log('투표 결과 불러오기 실패 : ', error);
-        onError?.(error);
-      } finally {
-        setResultsLoading(false);
-      }
-    };
-
-    fetchResultsByPollIds();
-  }, [pollIds, onError]);
-
-  // 댓글 불러오기
-  useEffect(() => {
-    const getAllComments = async () => {
-      try {
-        setCommentsLoading(true);
-        const res = await getComments();
-        setComments(res);
-      } catch (error) {
-        console.log('댓글 불러오기 실패 : ', error);
-        onError?.(error);
-      } finally {
-        setCommentsLoading(false);
-      }
-    };
-    getAllComments();
-  }, [onError]);
+  const {
+    results,
+    comments,
+    commentText,
+    setCommentText,
+    resultsLoading,
+    commentsLoading,
+    commentsEndRef,
+    submitComment,
+    refreshComments,
+  } = useComment({ onError });
 
   // isLoading 값은 두 로딩 상태가 모두 끝났을 때 false
   useEffect(() => {
-    console.log('resultsLoading:', resultsLoading, 'commentsLoading:', commentsLoading);
-    if (!resultsLoading && !commentsLoading) {
-      setIsLoading(false);
-    } else {
-      setIsLoading(true);
-    }
-  }, [resultsLoading, commentsLoading]);
+    setIsLoading(resultsLoading || commentsLoading);
+  }, [resultsLoading, commentsLoading, setIsLoading]);
 
-  // 댓글 등록 후 스크롤 이동 호출
-  const scrollToBottom = () => {
-    commentsEndRef.current?.scrollIntoView({ behavior: 'auto' });
-  };
-
-  // 댓글 목록이 바뀔 때마다 스크롤을 아래로
+  // 댓글 목록이 바뀔 때마다 스크롤 아래로
   useEffect(() => {
     if (comments.length > 0) {
       requestAnimationFrame(() => {
         commentsEndRef.current?.scrollIntoView({ behavior: 'auto' });
       });
     }
-  }, [comments]);
+  }, [comments, commentsEndRef]);
 
   // 리사이즈 이벤트 감지 후 스크롤 내리기
   useEffect(() => {
     const handleResize = () => {
-      // 댓글이 존재하면 맨 아래로 스크롤
       if (comments.length > 0) {
         commentsEndRef.current?.scrollIntoView({ behavior: 'auto' });
       }
     };
 
     window.addEventListener('resize', handleResize);
-    return () => {
-      window.removeEventListener('resize', handleResize);
-    };
-  }, [comments]);
+    return () => window.removeEventListener('resize', handleResize);
+  }, [comments, commentsEndRef]);
 
-  // 댓글 등록
-  const handleSubmitComment = async () => {
-    if (!commentText.trim()) {
-      alert('내용을 입력해주세요');
-      return;
-    }
-    try {
-      await postComment({
-        user_id: user.userId,
-        comment_text: commentText,
-      });
-
-      const updatedComments = await getComments();
-      setComments(updatedComments);
-      setCommentText('');
-      scrollToBottom();
-    } catch (error) {
-      console.log('댓글 등록 실패', error);
-      onError?.(error);
-    }
-  };
-
-  // 댓글 목록 새로고침 함수
-  const refreshComments = async () => {
-    try {
-      const res = await getComments();
-      setComments(res);
-    } catch (error) {
-      console.error('댓글 갱신 실패', error);
-      onError?.(error);
-    }
-  };
+  if (resultsLoading || commentsLoading) {
+    return (
+      <div className="flex justify-center items-center h-full w-full">
+        <Loading />
+      </div>
+    );
+  }
 
   return (
     <div className="md:h-full flex flex-col justify-center items-center">
@@ -200,7 +94,7 @@ const CommentContext = ({ setIsLoading, onError }: CommentProps) => {
           </div>
         </div>
 
-        {/* 댓글 영역 - 로딩 중에는 아예 렌더링하지 않음 */}
+        {/* 댓글 영역 */}
         <div className="h-[550px] md:h-[550px] border border-gray-300 rounded-2xl flex flex-col justify-center items-center px-4 py-2">
           {/* 댓글 리스트 */}
           <div className="flex-1 overflow-y-auto pr-2">
@@ -230,7 +124,7 @@ const CommentContext = ({ setIsLoading, onError }: CommentProps) => {
               nickname={user.randomNickname}
               comment={commentText}
               onChangeComment={(e) => setCommentText(e.target.value)}
-              onSubmit={handleSubmitComment}
+              onSubmit={() => submitComment(user.userId)}
             />
           </div>
         </div>
